backend/requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
azure-cognitiveservices-speech
azure-core==1.38.1
azure-cosmos
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
fastapi==0.128.7
ffmpy==1.0.0
filelock==3.20.3
fsspec==2026.2.0
google-auth==2.48.0
google-genai==1.62.0
gradio==6.5.1
gradio_client==2.0.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.2
openai==2.20.0
orjson==3.11.7
packaging==25.0
pandas==3.0.0
pillow==12.1.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
# PyAudio on Windows should be installed with conda-forge to avoid portaudio.h build errors
PyAudio==0.2.11; platform_system != 'Windows'
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
rsa==4.9.1
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
tomlkit==0.13.3
tqdm==4.67.3
typer==0.21.2
typer-slim==0.21.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
websockets==15.0.1
google-api-python-client==2.162.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
